+++ src/Form/ProjetParticipantType.php
@@ -3,7 +3,8 @@
 namespace App\Form;
 
 use App\Entity\ProjetParticipant;
-use App\Entity\Societe;
+use App\Form\Custom\ParticipantRoleChoiceType;
+use App\Form\Custom\SameSocieteUserType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;